% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_mds.R
\name{fast_mds}
\alias{fast_mds}
\title{Fast MDS}
\usage{
fast_mds(x, l, s, k)
}
\arguments{
\item{x}{Data matrix.}

\item{l}{The highest value where classical MDS can be computed efficiently.}

\item{s}{Number of sampling points. It should be 2 x estimated data dimension.}

\item{k}{Number of principal coordinates.}
}
\value{
Returns MDS based on Fast MDS algorithm as well as the first k eigenvalues.
\describe{
\item{points}{MDS}
\item{eigen}{eigenvalues}
}
}
\description{
Perfroms Multidimensional Scaling based on Yang, Tynia & Liu, Jinze & Mcmillan, Leonard & Wang, Wei. (2006).
}
