% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower_interpolation_mds.R
\name{gower_interpolation_mds}
\alias{gower_interpolation_mds}
\title{MDS based on Gower interpolation formula}
\usage{
gower_interpolation_mds(x, l, k)
}
\arguments{
\item{x}{Data matrix.}

\item{l}{The largest value which allows classical MDS to be computed efficiently, i.e, the larges value which makes
\code{cmdscale()} be run without any computational issues.}

\item{k}{Number of principal coordinates.}
}
\value{
Returns a list containing the following elements:
\describe{
\item{points}{A matrix that consists of \emph{k} columns corresponding to the MDS coordinates.}
\item{eigen}{The first \emph{k} eigenvalues.}
}
}
\description{
Performs \emph{Multidimensional Scaling} for big datasets. This method can compute a MDS configuration
even when the dataset is so large that classical MDS methods (\code{cmdscale}) can not be run due to computational
problems.
}
\details{
Gower interpolation formula allows to add a new set of points to a given MDS configuration.

Given a matrix \emph{X nxp}, a MDS configuration for this matrix, and a matrix \emph{X_new mxp}, one wants to add these
new \emph{m} rows to the existing MDS configuration. So,after adding these new rows,the MDS configuration will have \emph{n+m}
rows.
}
\examples{
x <- matrix(data = rnorm(4*10000, sd = 10), nrow = 10000)
cmds <- gower_interpolation_mds(x = x, l = 100, k = 2)
head(cmds$points)
cmds$eigen
}
\references{
Gower, J.C. and D.J, Hand (1995). \emph{Biplots}, Volume 54. CRC Press.

Borg and Groenen (1997). \emph{Modern Multidimensional Scaling}. New York: Springer. pp. 340-342.
}
